requirements.txt
streamlit
pdfminer.six
reportlab
